nuevo modelo de bd, mobiliario y puntera

+++ PlanVisitaWebAPI/Models/Sucursal/ClientesHBFDataSetAttribute.cs
@@ -12,7 +12,7 @@ namespace PlanVisitaWebAPI.Models.Sucursal
         public string Address { get; set; }
         public string city { get; set; }
         public string street { get; set; }
-        public Nullable<short> GroupCode { get; set; }
+        public Nullable<int> GroupCode { get; set; }
         public string GroupName { get; set; }
         public string Address2 { get; set; }
         public string Estado { get; set; }
